Add mutation for updating preferred name
